[gui] functioning gui able to load and view raw and whitened traces

+++ kilosort/gui/message_log_box.py
@@ -40,12 +40,13 @@ class MessageLogBox(QtWidgets.QGroupBox):
         context = self.get_context()
         if context is not None:
             context_path = context.context_path
-            log_file_name = context.raw_data.name + ".log"
+            log_file_name = context.data_path.stem + ".log"
 
             log_file_path = Path(context_path) / log_file_name
 
-            with open(log_file_path, "w") as log_file:
-                log_file.write(self.log_box.toPlainText())
+            if log_file_path.exists():
+                with open(log_file_path, "w") as log_file:
+                    log_file.write(self.log_box.toPlainText())
 
     def get_context(self):
         return self.gui.context
